# MasterClawInterface/openapi.yaml
openapi: 3.0.3
info:
  title: MasterClaw Chat Interface API
  description: |
    Real-time WebSocket and REST API for Rex to communicate with AI agents.
    
    ## Features
    - **WebSocket Messaging**: Real-time bidirectional communication
    - **Agent Registry**: Manage agent registration and status
    - **Message Routing**: Direct and broadcast messaging
    - **Memory Logging**: Log agent thoughts, needs, observations
    - **Job Management**: Create and track agent tasks
    
    ## Authentication
    API uses JWT Bearer tokens for authentication. Include token in header:
    ```
    Authorization: Bearer <token>
    ```
  version: 1.0.0
  contact:
    name: MasterClaw Team
    url: https://github.com/TheMasterClaw/masterclaw-core
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.masterclaw.io
    description: Production server
...
paths:
...
  /api/messages/conversation/{agent1}/{agent2}:
    get:
      tags:
        - Messages
      summary: Get conversation
      description: Get message history between two agents
      parameters:
        - name: agent1
          in: path
          required: true
          schema:
            type: string
        - name: agent2
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Conversation messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
  /api/messages/agent/{agentId}:
    get:
      tags:
        - Messages
      summary: Get agent messages
      description: Get all messages for a specific agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Agent messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'

    post:
      tags:
        - Messages
      summary: Send message (REST)
      description: Send a message via REST API (fallback when WebSocket unavailable)
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fromAgent
                - toAgent
                - content
              properties:
                fromAgent:
                  type: string
                  description: Sender agent ID
                toAgent:
                  type: string
                  description: Recipient agent ID
                content:
                  type: string
                  description: Message content
                type:
                  type: string
                  enum: [text, command, status, memory, error, system]
                  default: text
                metadata:
                  type: object
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    $ref: '#/components/schemas/Message'
  /api/messages/agent/{agentId}/unread:
    get:
      tags:
        - Messages
      summary: Get unread count
      description: Get the number of unread messages for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unread message count
          content:
            application/json:
              schema:
                type: object
                properties:
                  agentId:
                    type: string
                  unreadCount:
                    type: integer
  /api/messages/{messageId}/read:
    put:
      tags:
        - Messages
      summary: Mark message as read
      description: Mark a specific message as read
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    $ref: '#/components/schemas/Message'
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        messageId:
          type: string
        fromAgent:
          type: string
        toAgent:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [text, command, status, memory, error, system]
        metadata:
          type: object
        delivered:
          type: boolean
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
...